+++ src/init.c
@@ -11,6 +11,8 @@
 extern SEXP _vapour_raster_info_cpp(SEXP);
 extern SEXP _vapour_raster_io_cpp(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _vapour_sds_info_cpp(SEXP);
+extern SEXP _vapour_vapour_all_drivers_cpp();
+extern SEXP _vapour_vapour_gdal_version_cpp();
 extern SEXP _vapour_vapour_layer_names_cpp(SEXP, SEXP);
 extern SEXP _vapour_vapour_projection_info_cpp(SEXP, SEXP, SEXP);
 extern SEXP _vapour_vapour_read_attributes_cpp(SEXP, SEXP, SEXP, SEXP);
@@ -21,6 +23,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_vapour_raster_info_cpp",            (DL_FUNC) &_vapour_raster_info_cpp,            1},
     {"_vapour_raster_io_cpp",              (DL_FUNC) &_vapour_raster_io_cpp,              4},
     {"_vapour_sds_info_cpp",               (DL_FUNC) &_vapour_sds_info_cpp,               1},
+    {"_vapour_vapour_all_drivers_cpp",     (DL_FUNC) &_vapour_vapour_all_drivers_cpp,     0},
+    {"_vapour_vapour_gdal_version_cpp",    (DL_FUNC) &_vapour_vapour_gdal_version_cpp,    0},
     {"_vapour_vapour_layer_names_cpp",     (DL_FUNC) &_vapour_vapour_layer_names_cpp,     2},
     {"_vapour_vapour_projection_info_cpp", (DL_FUNC) &_vapour_vapour_projection_info_cpp, 3},
     {"_vapour_vapour_read_attributes_cpp", (DL_FUNC) &_vapour_vapour_read_attributes_cpp, 4},