+++ NEWS.md
@@ -1,5 +1,7 @@
 # vapour 0.2.0
 
+* New functions `vapour_gdal_version` and `vapour_all_drivers` to return information about the GDAL library in use. 
+
 * Function `vapour_read_names` is now implemented using the library API, rather than via OGRSQL. 
 
 * New function `vapour_geom_summary` to return basic identity, validity, and extent of feature geometries.